add a model folder for extras and move sprungbrett
