+++ Integreat/Integreat.Shared/Models/Extras/Sprungbrett/SprungbrettPager.cs
@@ -1,0 +1,15 @@
+﻿using Newtonsoft.Json;
+
+namespace Integreat.Shared.Models.Extras.Sprungbrett
+{
+    /// <summary>
+    /// additional pager information parsed from json
+    /// </summary>
+    public class SprungbrettPager
+    {
+        [JsonProperty("current")]
+        public int Current { get; set; }
+        [JsonProperty("max")]
+        public int Max { get; set; }
+    }
+}
