--- Integreat/Integreat.Shared/Models/Sprungbrett/SprungbrettPager.cs
@@ -1,14 +1,0 @@
-﻿using Newtonsoft.Json;
-
-namespace Integreat.Shared.Models.Sprungbrett
-{
-
-        public class SprungbrettPager
-        {
-            [JsonProperty("current")]
-            public int Current { get; set; }
-            [JsonProperty("max")]
-            public int Max { get; set; }
-        }
-    
-}
